split off cost chart specs from cost.py
