+++ src/actions/reports/cost.py
@@ -2,10 +2,7 @@ import html
 import inspect
 import numpy as np
 import re
-from collections import namedtuple
-from collections.abc import Callable, Iterable, Mapping
-from copy import deepcopy
-from dataclasses import dataclass, field
+from collections.abc import Iterable
 from operator import attrgetter
 
 from matplotlib import pyplot as plt
@@ -16,6 +13,7 @@ from objects import PlanPrinter
 from objects import PlanNode
 from actions.report import AbstractReportAction
 from actions.reports.cost_metrics import CostMetrics
+from actions.reports.cost_chart_specs import CostChartSpecs, ChartSpec, DataPoint, PlotType
 
 
 BOXPLOT_DESCRIPTION = (
@@ -51,86 +49,6 @@ X_TIME_COST_CHART_DESCRIPTION = (
 )
 
 
-DataPoint = namedtuple('DataPoint', ['x', 'cost', 'time_ms', 'node'])
-
-
-@dataclass(frozen=True)
-class ChartOptions:
-    adjust_cost_by_actual_rows: bool = True
-    multipy_by_nloops: bool = False
-    log_scale_x: bool = False
-    log_scale_cost: bool = False
-    log_scale_time: bool = False
-
-    bp_show_fliers: bool = True  # boxplot only
-
-    def __str__(self):
-        return ','.join(filter(lambda a: getattr(self, a), self.__dict__.keys()))
-
-
-@dataclass
-class ChartSpec:
-    plotter: Callable[['ChartSpec'], bool]
-    title: str
-    description: str
-    xlabel: str
-    ylabel1: str
-    ylabel2: str
-    query_filter: Callable[[str], bool]
-    node_filter: Callable[[PlanNode], bool]
-    x_getter: Callable
-    series_suffix: Callable = lambda node: ''
-    options: ChartOptions = field(default_factory=ChartOptions)
-
-    xtra_query_filter_list: Iterable[Callable[[str], bool]] = field(default_factory=list)
-    xtra_node_filter_list: Iterable[Callable[[PlanNode], bool]] = field(default_factory=list)
-
-    file_name: str = ''
-    queries: set[str] = field(default_factory=set)
-    series_data: Mapping[str: Iterable[DataPoint]] = field(default_factory=dict)
-    series_format: Mapping[str: str] = field(default_factory=dict)
-    outliers: Mapping[str: Iterable[DataPoint]] = field(default_factory=dict)
-    outlier_axis: str = 'cost/time ratio'
-
-    def is_boxplot(self):
-        return self.plotter is CostReport.draw_boxplot
-
-    def make_variant(self, xtra_title, overwrite_title=False,
-                     description: str = None,
-                     xtra_query_filter: Callable[[str], bool] = None,
-                     xtra_node_filter: Callable[[PlanNode], bool] = None,
-                     xlabel: str = None,
-                     x_getter: Callable = None,
-                     series_suffix: str = None,
-                     options: ChartOptions = None):
-        var = deepcopy(self)
-        if overwrite_title:
-            var.title = xtra_title
-        else:
-            var.title += f' ({xtra_title})'
-        if description:
-            var.description = description
-        if xtra_query_filter:
-            var.xtra_query_filter_list.append(xtra_query_filter)
-        if xtra_node_filter:
-            var.xtra_node_filter_list.append(xtra_node_filter)
-        if xlabel:
-            var.xlabel = xlabel
-        if x_getter:
-            var.x_getter = x_getter
-        if series_suffix:
-            var.series_suffix = series_suffix
-        if options:
-            var.options = options
-        return var
-
-    def test_query(self, query_str):
-        return all(f(query_str) for f in [self.query_filter, *self.xtra_query_filter_list])
-
-    def test_node(self, node):
-        return all(f(node) for f in [self.node_filter, *self.xtra_node_filter_list])
-
-
 class CostReport(AbstractReportAction):
     def __init__(self):
         super().__init__()
@@ -153,8 +71,8 @@ class CostReport(AbstractReportAction):
 
         chart_specs = list()
         if interactive:
-            chart_specs = report.choose_chart_spec(report.get_xtc_chart_specs(cm)
-                                                   + report.get_exp_chart_specs(cm))
+            chart_specs = report.choose_chart_spec(CostChartSpecs.get_xtc_chart_specs(cm)
+                                                   + CostChartSpecs.get_exp_chart_specs(cm))
         else:
             report.define_version(loq.db_version)
             report.report_config(loq.config, "YB")
@@ -179,10 +97,10 @@ class CostReport(AbstractReportAction):
         if interactive:
             report.collect_nodes_and_create_plots(chart_specs)
         else:
-            dist_chart_specs = report.get_dist_chart_specs(cm)
-            xtc_chart_specs = report.get_xtc_chart_specs(cm)
-            exp_chart_specs = report.get_exp_chart_specs(cm)
-            # exp_chart_specs += report.get_more_exp_chart_specs(cm)
+            dist_chart_specs = CostChartSpecs.get_dist_chart_specs(cm)
+            xtc_chart_specs = CostChartSpecs.get_xtc_chart_specs(cm)
+            exp_chart_specs = CostChartSpecs.get_exp_chart_specs(cm)
+            # exp_chart_specs += CostChartSpecs.get_more_exp_chart_specs(cm)
             report.collect_nodes_and_create_plots(dist_chart_specs
                                                   + xtc_chart_specs + exp_chart_specs)
             report.build_report(dist_chart_specs, xtc_chart_specs, exp_chart_specs)
@@ -196,44 +114,44 @@ class CostReport(AbstractReportAction):
 
     def build_report(self, dist_chart_specs, xtc_chart_specs, exp_chart_specs):
         def report_one_chart(self, id, spec):
-            self.content  += f"=== {id}. {html.escape(spec.title)}\n{spec.description}\n"
+            self.content += f"=== {id}. {html.escape(spec.title)}\n{spec.description}\n"
             self.report_chart(spec)
 
         id = 0
-        self.content  += "\n== Time & Cost <<_boxplot_chart, Distribution Charts>>\n"
+        self.content += "\n== Time & Cost <<_boxplot_chart, Distribution Charts>>\n"
         for spec in dist_chart_specs:
             report_one_chart(self, id, spec)
             id += 1
 
-        self.content  += "\n== <<_time_cost_relationship_charts, Time - Cost Relationship Charts>>\n"
+        self.content += "\n== <<_time_cost_relationship_charts, Time - Cost Relationship Charts>>\n"
         for spec in xtc_chart_specs:
             report_one_chart(self, id, spec)
             id += 1
 
-        self.content  += "\n== Experimental Charts\n"
+        self.content += "\n== Experimental Charts\n"
         for spec in exp_chart_specs:
             report_one_chart(self, id, spec)
             id += 1
 
-        self.content  += "== Chart Descriptions\n"
-        self.content  += BOXPLOT_DESCRIPTION
-        self.content  += X_TIME_COST_CHART_DESCRIPTION
+        self.content += "== Chart Descriptions\n"
+        self.content += BOXPLOT_DESCRIPTION
+        self.content += X_TIME_COST_CHART_DESCRIPTION
 
     def report_chart_filters(self, spec: ChartSpec):
         self.start_collapsible("Chart specifications")
         self.start_source(["python"])
-        self.content  += "=== Query Filters ===\n"
+        self.content += "=== Query Filters ===\n"
         for f in spec.query_filter, *spec.xtra_query_filter_list:
-            self.content  += inspect.getsource(f)
-        self.content  += "=== Node Filters ===\n"
+            self.content += inspect.getsource(f)
+        self.content += "=== Node Filters ===\n"
         for f in spec.node_filter, *spec.xtra_node_filter_list:
-            self.content  += inspect.getsource(f)
-        self.content  += "=== X Axsis Data ===\n"
-        self.content  += inspect.getsource(spec.x_getter)
-        self.content  += "=== Series Suffix ===\n"
-        self.content  += inspect.getsource(spec.series_suffix)
-        self.content  += "=== Options ===\n"
-        self.content  += str(spec.options)
+            self.content += inspect.getsource(f)
+        self.content += "=== X Axsis Data ===\n"
+        self.content += inspect.getsource(spec.x_getter)
+        self.content += "=== Series Suffix ===\n"
+        self.content += inspect.getsource(spec.series_suffix)
+        self.content += "=== Options ===\n"
+        self.content += str(spec.options)
         self.end_source()
         self.end_collapsible()
 
@@ -262,13 +180,13 @@ class CostReport(AbstractReportAction):
             self.content += table_header
             self.end_table_row()
             for x, cost, time_ms, node in data_points:
-                self.content  += f">|{x:.3f}\n>|{time_ms:.3f}\n>|{cost:.3f}\n|\n"
+                self.content += f">|{x:.3f}\n>|{time_ms:.3f}\n>|{cost:.3f}\n|\n"
                 self.start_source(["sql"], linenums=False)
-                self.content  += str(node)
+                self.content += str(node)
                 self.end_source()
-                self.content  += "|\n"
+                self.content += "|\n"
                 self.start_source(["sql"], linenums=False)
-                self.content  += cm.get_node_query(node).query
+                self.content += cm.get_node_query(node).query
                 self.end_source()
 
             self.end_table()
@@ -353,7 +271,7 @@ class CostReport(AbstractReportAction):
     def make_file_name(self, str_list: Iterable[str]):
         return f"{'-'.join(s.strip(self.__spcrs).translate(self.__xtab) for s in str_list)}.svg"
 
-    def draw_x_cost_time_plot(self, spec):
+    def draw_x_time_cost_plot(self, spec):
         title = spec.title
         xy_labels = [spec.xlabel, spec.ylabel1, spec.ylabel2]
 
@@ -551,6 +469,11 @@ class CostReport(AbstractReportAction):
                         'h', 'H', 'D', 'd', 'P', 'X']
         line_style = ['-', '--', '-.', ':']
 
+        plotters = {
+            PlotType.BOXPLOT: CostReport.draw_boxplot,
+            PlotType.X_TIME_COST_PLOT: CostReport.draw_x_time_cost_plot,
+        }
+
         for spec in specs:
             for i, (series_label, data_points) in enumerate(sorted(spec.series_data.items())):
                 fmt = self.get_series_color(series_label)
@@ -558,7 +481,7 @@ class CostReport(AbstractReportAction):
                 fmt += line_style[(i+5) % len(line_style)]
                 spec.series_format[series_label] = fmt
 
-            spec.plotter(self, spec)
+            plotters[spec.plotter](self, spec)
 
     def choose_chart_spec(self, chart_specs):
         choices = '\n'.join([f'{n}: {s.title}' for n, s in enumerate(chart_specs)])
@@ -609,8 +532,8 @@ class CostReport(AbstractReportAction):
 
             modifiers = event.mouseevent.modifiers
             if 'alt' in modifiers:
-                ptree = self.get_node_plan_tree(node)
-                ann.set_text(f'{PlanPrinter.build_plan_tree_str(ptree)}')
+                ptree = self.cm.get_node_plan_tree(node)
+                ann.set_text(PlanPrinter.build_plan_tree_str(ptree))
             elif 'shift' in modifiers:
                 query = self.cm.get_node_query(node)
                 ann.set_text(f'{query.query_hash}\n{query.query}')
@@ -654,376 +577,3 @@ class CostReport(AbstractReportAction):
         fig.canvas.mpl_connect('pick_event', on_pick)
         fig.canvas.mpl_connect('button_release_event', on_button_release)
         plt.show()
-
-    def get_dist_chart_specs(self, cm: CostMetrics):
-        return [
-            (boxplot_simple_scan_node := ChartSpec(
-                CostReport.draw_boxplot,
-                '',
-                ('* Nodes with local filter, recheck-removed-rows or partial aggregate'
-                 ' are _excluded_\n'),
-                'xlabel', 'Node type', '',
-                lambda query: True,
-                lambda node: (
-                    float(node.nloops) == 1
-                    and float(node.rows) >= 1
-                    and cm.get_node_width(node) == 0
-                    and cm.has_no_local_filtering(node)
-                    and not cm.has_partial_aggregate(node)
-                ),
-                x_getter=lambda node: 1,
-            )).make_variant(
-                'Per row cost/time ratio of the scan nodes (width=0, rows>=1)', True,
-                ('  ((total_cost - startup_cost) / estimated_rows)'
-                 ' / ((total_time - startup_time) / actual_rows)\n'
-                 '\n* Nodes with local filter, recheck-removed-rows or partial aggregate'
-                 ' are _excluded_\n'),
-                xlabel='Per row cost/time ratio [1/ms]',
-                x_getter=lambda node: (cm.get_per_row_cost(node)
-                                       / (cm.get_per_row_time(node) or 0.01)),
-            ),
-            boxplot_simple_scan_node.make_variant(
-                'Per row time of the scan nodes (width=0, rows>=1)', True,
-                xlabel='Per row execution time[ms]',
-                x_getter=lambda node: cm.get_per_row_time(node),
-            ),
-            boxplot_simple_scan_node.make_variant(
-                'Per row cost of the scan nodes (width=0, rows>=1)', True,
-                xlabel='Per row cost',
-                x_getter=lambda node: cm.get_per_row_cost(node),
-            ),
-            boxplot_simple_scan_node.make_variant(
-                'Startup cost/time ratio of the scan nodes (width=0)', True,
-                xlabel='Startup cost/time ratio [1/ms]',
-                x_getter=lambda node: (float(node.startup_cost)
-                                       / (float(node.startup_ms)
-                                          if float(node.startup_ms) else 0.001)),
-            ),
-            boxplot_simple_scan_node.make_variant(
-                'Startup time of the scan nodes (width=0)', True,
-                xlabel='Startup time [ms]',
-                x_getter=lambda node: float(node.startup_ms),
-            ),
-        ]
-
-    def get_xtc_chart_specs(self, cm: CostMetrics):
-        return (self.get_column_and_value_metric_chart_specs(cm)
-                + self.get_simple_index_scan_chart_specs(cm)
-                + self.get_in_list_chart_specs(cm))
-
-    def get_exp_chart_specs(self, cm: CostMetrics):
-        return self.get_composite_key_access_chart_specs(cm)
-
-    def get_simple_index_scan_chart_specs(self, cm: CostMetrics):
-        return [
-            (chart_simple_index_scan := ChartSpec(
-                CostReport.draw_x_cost_time_plot,
-                ('Simple index access conditions and corresponding seq scans by node type'),
-                ('Index (Only) Scans with simple index access condition on single key item'
-                 ' and the Seq Scans from the same queries.'
-                 '\n\n* No IN-list, OR\'ed condition, etc.'
-                 '\n\n* The nodes showing "Rows Removed by (Index) Recheck" are excluded.'
-                 '\n\n* No nodes from EXISTS and JOIN queries'),
-                'Output row count', 'Estimated cost', 'Execution time [ms]',
-                lambda query: (
-                    cm.is_single_table_query(query)
-                    and cm.has_no_filter_indexscan(query)
-                    and not cm.has_local_filter(query)
-                    and not cm.has_aggregate(query)
-                ),
-                lambda node: (
-                    cm.has_no_local_filtering(node)
-                    and not cm.has_no_condition(node)
-                    and (node.is_seq_scan
-                         or (node.is_any_index_scan
-                             and cm.has_only_simple_condition(node, index_cond_only=True)))
-                ),
-                x_getter=lambda node: float(node.rows),
-                series_suffix=(lambda node:
-                               f'{node.index_name or node.table_name}:'
-                               f'width={cm.get_node_width(node)}'),
-            )).make_variant(
-                't100000 and t100000w',
-                xtra_query_filter=lambda query: 't100000 ' in query or 't100000w ' in query,
-                xtra_node_filter=(lambda node:
-                                  float(cm.get_table_row_count(node.table_name)) == 100000),
-                series_suffix=(lambda node:
-                               f'{node.table_name}:width={cm.get_node_width(node)}'),
-            ),
-            chart_simple_index_scan.make_variant(
-                't100000',
-                xtra_query_filter=lambda query: 't100000 ' in query,
-                xtra_node_filter=(lambda node:
-                                  float(cm.get_table_row_count(node.table_name)) == 100000),
-            ),
-            chart_simple_index_scan.make_variant(
-                't100000w',
-                xtra_query_filter=lambda query: 't100000w ' in query,
-                xtra_node_filter=(lambda node:
-                                  float(cm.get_table_row_count(node.table_name)) == 100000),
-            ),
-            chart_simple_index_scan.make_variant(
-                't10000',
-                xtra_query_filter=lambda query: 't10000 ' in query,
-                xtra_node_filter=(lambda node:
-                                  float(cm.get_table_row_count(node.table_name)) == 10000),
-            ),
-            chart_simple_index_scan.make_variant(
-                't1000',
-                xtra_query_filter=lambda query: 't1000 ' in query,
-                xtra_node_filter=(lambda node:
-                                  float(cm.get_table_row_count(node.table_name)) == 1000),
-            ),
-            chart_simple_index_scan.make_variant(
-                't100',
-                xtra_query_filter=lambda query: 't100 ' in query,
-                xtra_node_filter=(lambda node:
-                                  float(cm.get_table_row_count(node.table_name)) == 100),
-            ),
-        ]
-
-    def get_column_and_value_metric_chart_specs(self, cm: CostMetrics):
-        column_and_value_metric_chart = ChartSpec(
-            CostReport.draw_x_cost_time_plot,
-            'Column/value num/position metric test queries', '',
-            '', 'Estimated cost', 'Execution time [ms]',
-            lambda query: (
-                cm.is_single_table_query(query)
-                and not cm.has_aggregate(query)
-            ),
-            lambda node: (
-                node.table_name == 't1000000c10'
-                and not cm.has_partial_aggregate(node)
-            ),
-            x_getter=lambda node: 0,
-        )
-
-        column_and_value_metric_single_column_chart = (
-            column_and_value_metric_chart.make_variant(
-                'Column/value num/position metric test queries (single column)', '',
-                xtra_query_filter=lambda query: (
-                    len(cm.get_columns_in_query(query)) == 1
-                ),
-            )
-        )
-
-        return [
-            column_and_value_metric_chart.make_variant(
-                'Column count (select-list, no condition)', True,
-                xtra_node_filter=lambda node: (
-                    cm.has_no_condition(node)
-                    and (len(cm.get_columns_in_query(cm.get_node_query_str(node))) > 1
-                         or (cm.get_columns_in_query(cm.get_node_query_str(node))
-                             in (['c4'], ['c5'])))
-                ),
-                xlabel='Column count',
-                x_getter=lambda node: (
-                    len(cm.get_columns_in_query(cm.get_node_query_str(node)))
-                ),
-            ),
-            column_and_value_metric_chart.make_variant(
-                'Column count (columns in condition)', True,
-                xtra_query_filter=lambda query: (
-                    re.match(r'select 0 from t1000000c10 where [c0-9+ ]+ = +500000', query)
-                ),
-                xlabel='Column count',
-                x_getter=lambda node: (
-                    len(cm.get_columns_in_query(cm.get_node_query_str(node)))
-                ),
-                series_suffix=lambda node: (
-                    ''.join([
-                        '(remote filter)' if cm.has_only_scan_filter_condition(node) else ''
-                    ])),
-            ),
-            column_and_value_metric_single_column_chart.make_variant(
-                'Column position (select-list, no condition)', True,
-                xtra_node_filter=lambda node: (
-                    cm.has_no_condition(node)
-                    and cm.get_node_width(node) == 4
-                ),
-                xlabel='Column position',
-                x_getter=lambda node: cm.get_single_column_query_column_position(node),
-            ),
-            column_and_value_metric_single_column_chart.make_variant(
-                'Column position (select-list and condition)', True,
-                xtra_node_filter=lambda node: (
-                    not cm.has_no_condition(node)
-                    and cm.has_only_simple_condition(node,
-                                                     index_cond_only=False,
-                                                     index_key_prefix_only=True)
-                    and abs(0.5 - cm.get_single_column_node_normalized_eq_cond_value(node)) < 0.01
-                    and cm.get_node_width(node) == 4
-                ),
-                xlabel='Column position',
-                x_getter=lambda node: cm.get_single_column_query_column_position(node),
-            ),
-            column_and_value_metric_single_column_chart.make_variant(
-                'Normalizd value (value position) in condition', True,
-                xtra_node_filter=lambda node: (
-                    not cm.has_no_condition(node)
-                    and cm.has_only_simple_condition(node,
-                                                     index_cond_only=True,
-                                                     index_key_prefix_only=True)
-                    and cm.get_single_column_node_normalized_eq_cond_value(node) is not None
-                    and cm.get_node_width(node) == 4
-                ),
-                xlabel='Normalized value',
-                x_getter=lambda node: cm.get_single_column_node_normalized_eq_cond_value(node),
-                series_suffix=lambda node: node.index_name or '',
-            ),
-        ]
-
-    def get_in_list_chart_specs(self, cm: CostMetrics):
-        return [
-            chart_literal_in_list := ChartSpec(
-                CostReport.draw_x_cost_time_plot,
-                'Index scan nodes with literal IN-list',
-                ("Index (Only) Scans with literal IN-list in the index access condition."),
-                'Output row count', 'Estimated cost', 'Execution time [ms]',
-                lambda query: True,
-                lambda node: (
-                    cm.has_literal_inlist_index_cond(node)
-                    and cm.has_no_local_filtering(node)
-                ),
-                x_getter=lambda node: float(node.rows),
-                series_suffix=(lambda node:
-                               ''.join([
-                                   f'{node.index_name or node.table_name}:',
-                                   f'width={cm.get_node_width(node)}',
-                                   ' ncInItems=',
-                                   cm.build_non_contiguous_literal_inlist_count_str(
-                                       node.table_name, node.get_index_cond()),
-                               ])),
-            ),
-            chart_literal_in_list.make_variant(
-                "output <= 200 rows",
-                xtra_node_filter=lambda node: float(node.rows) <= 100,
-            ),
-            ChartSpec(
-                CostReport.draw_x_cost_time_plot,
-                'Parameterized IN-list index scans (BNL)',
-                ("Index (Only) Scans with BNL-generaed parameterized IN-list, plus the Seq Scans"
-                 " from the same queries."),
-                'Output row count', 'Estimated cost', 'Execution time [ms]',
-                lambda query: True,
-                lambda node: (
-                    cm.has_bnl_inlist_index_cond(node)
-                    and cm.has_no_local_filtering(node)
-                ),
-                x_getter=lambda node: float(node.rows),
-                series_suffix=(lambda node:
-                               f'{node.index_name}:width={cm.get_node_width(node)}'
-                               f' loops={node.nloops}'
-                               ),
-            ),
-        ]
-
-    def get_composite_key_access_chart_specs(self, cm: CostMetrics):
-        return [
-            chart_composite_key := ChartSpec(
-                CostReport.draw_x_cost_time_plot,
-                'Composite key index scans',
-                ("* The clustered plots near the lower left corner need adjustments the series"
-                 " criteria and/or node filtering."),
-                'Output row count', 'Estimated cost', 'Execution time [ms]',
-                lambda query: 't1000000m' in query or 't1000000c10' in query,
-                lambda node: (
-                    cm.has_no_local_filtering(node)
-                    and cm.has_only_simple_condition(node, index_cond_only=True)
-                ),
-                x_getter=lambda node: float(node.rows),
-                series_suffix=lambda node: f'{node.index_name}',
-            ),
-            chart_composite_key.make_variant(
-                'output rows <= 100',
-                xtra_node_filter=lambda node: float(node.rows) <= 100,
-            ),
-            chart_composite_key.make_variant(
-                'output rows <= 100, x=output_row_count x key_prefix_gaps',
-                description=(
-                    "Index key prefix gaps: NDV of the keys before the first equality condition."
-                    "\n\ne.g.: for index key `(c3, c4, c5)`,"
-                    " condition: `c4 >= x and c5 = y` then the prefix NDV would be:"
-                    " `select count(*) from (select distinct c3, c4 from t where c4 >= x) v;`"),
-                xtra_node_filter=lambda node: float(node.rows) <= 100,
-                xlabel='Output row count x key prefix gaps',
-                x_getter=lambda node: float(node.rows) * cm.get_index_key_prefix_gaps(node),
-            ),
-            chart_composite_key.make_variant(
-                'key_prefix_gaps in series criteria',
-                description=(
-                    "Index key prefix gaps: NDV of the keys before the first equality condition."
-                    "\n\ne.g.: for index key `(c3, c4, c5)`,"
-                    " condition: `c4 >= x and c5 = y` then the prefix NDV would be:"
-                    " `select count(*) from (select distinct c3, c4 from t where c4 >= x) v;`"),
-                series_suffix=lambda node: (f'{node.index_name}'
-                                            ' gaps={cm.get_index_key_prefix_gaps(node)}'),
-            ),
-        ]
-
-    def get_more_exp_chart_specs(self, cm: CostMetrics):
-        return [
-            ChartSpec(
-                CostReport.draw_x_cost_time_plot,
-                'Scans with simple remote index and/or table filter',
-                "* Index (Only) Scans may or may not have index access condition as well.",
-                'Output row count', 'Estimated cost', 'Execution time [ms]',
-                lambda query: cm.has_scan_filter_indexscan(query),
-                lambda node: (
-                    cm.has_only_scan_filter_condition(node)
-                    or cm.has_only_simple_condition(node, index_cond_only=True)
-                ),
-                x_getter=lambda node: float(node.rows),
-                series_suffix=lambda node: (
-                    f'{node.index_name or node.table_name}'
-                    f':width={cm.get_node_width(node)} loops={node.nloops}'
-                ),
-            ),
-            ChartSpec(
-                CostReport.draw_x_cost_time_plot,
-                'Scans with remote filter(s)',
-                '',
-                'Output row count', 'Estimated cost', 'Execution time [ms]',
-                lambda query: True,
-                lambda node: cm.has_only_simple_condition(node, index_cond_only=False),
-                x_getter=lambda node: float(node.rows),
-                series_suffix=(lambda node:
-                               f'{node.index_name or node.table_name}'
-                               f':width={cm.get_node_width(node)} loops={node.nloops}'),
-            ),
-            chart_agg_pushdown := ChartSpec(
-                CostReport.draw_x_cost_time_plot,
-                'Full scan + agg push down by table rows',
-                ('Scan nodes from `select count(*) from ...` single table queries without'
-                 ' any search conditions'
-                 '\n\n* The costs are not adjusted'),
-                'Table rows', 'Estimated cost', 'Execution time [ms]',
-                lambda query: cm.has_aggregate(query),
-                lambda node: (
-                    cm.has_partial_aggregate(node)
-                    and cm.has_no_local_filtering(node)
-                    ),
-                x_getter=lambda node: float(cm.get_table_row_count(node.table_name)),
-                series_suffix=lambda node: f'{node.index_name or node.table_name}',
-                options=ChartOptions(adjust_cost_by_actual_rows=False),
-            ),
-            chart_agg_pushdown.make_variant(
-                'log scale',
-                options=ChartOptions(adjust_cost_by_actual_rows=False,
-                                     log_scale_x=True,
-                                     log_scale_cost=True,
-                                     log_scale_time=True),
-            ),
-            ChartSpec(
-                CostReport.draw_x_cost_time_plot,
-                'No filter full scans by output row x width',
-                '* need to adjust series grouping and query/node selection',
-                'Output rows x width', 'Estimated cost', 'Execution time [ms]',
-                lambda query: True,
-                lambda node: (node.has_no_filter()
-                              and not node.get_index_cond()
-                              and not cm.has_partial_aggregate(node)),
-                x_getter=lambda node: float(node.rows) * cm.get_node_width(node),
-                series_suffix=lambda node: f'{node.index_name or node.table_name}',
-            ),
-        ]
